add separate module for filter parsing
